+++ modules/fabric_1.20.1/src/main/java/com/armilp/ifreq/client/ClientMainEZ.java
@@ -1,13 +1,12 @@
 package com.armilp.ifreq.client;
 
-import com.armilp.ifreq.client.accessory.renderer.WalkieAccessoryRenderer;
 import com.armilp.ifreq.client.keys.ClientKeyHandler;
 import com.armilp.ifreq.client.keys.KeyBindings;
 import com.armilp.ifreq.client.screen.WalkieTalkieScreen;
-import com.armilp.ifreq.common.registry.ModItems;
 import com.armilp.ifreq.common.registry.ModMenus;
-import io.wispforest.accessories.api.client.AccessoriesRendererRegistry;
+import com.armilp.ifreq.client.compat.accessories.AccessoriesCompat;
 import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.loader.api.FabricLoader;
 import net.minecraft.client.gui.screen.ingame.HandledScreens;
 
 public class ClientMainEZ implements ClientModInitializer {
@@ -17,9 +16,6 @@ public class ClientMainEZ implements ClientModInitializer {
         KeyBindings.register();
         HandledScreens.register(ModMenus.WALKIE_TALKIE_MENU, WalkieTalkieScreen::new);
         ClientKeyHandler.register();
-        AccessoriesRendererRegistry.registerRenderer(
-                ModItems.WALKIE_TALKIE,
-                WalkieAccessoryRenderer::new
-        );
+        AccessoriesCompat.initClient();
     }
 }