+++ modules/fabric_1.20.1/changelog.md
@@ -1,3 +1,8 @@
-## 📦 Version 0.1-beta-fabric
+## 📦 Version 0.2-beta-fabric
 
-Initial release of Immersive Frequencies mod! This is a beta version, so expect some bugs and missing features.
+### Added
+  - Accessory slot for walkie-talkie
+  - 3d model for the walkie-talkie item
+  - walkie-talkie gui can be open while in blet slot of the Accessory slot hitting R key
+
+This is a beta version, so expect some bugs and missing features.